Move trace_exemplar to OC folder and simplify test
