--- opencensus/trace_exemplar_quickstart/main_test.go
@@ -1,35 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package trace_exemplar_quickstart
-
-import (
-	"bytes"
-	"strings"
-	"testing"
-
-	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
-)
-
-func TestWriteTimeSeriesData(t *testing.T) {
-	tc := testutil.SystemTest(t)
-	buf := &bytes.Buffer{}
-	if err := writeTimeSeriesData(buf, tc.ProjectID); err != nil {
-		t.Fatalf("writeTimeSeriesData: %v", err)
-	}
-	want := "Done"
-	if got := buf.String(); !strings.Contains(got, want) {
-		t.Fatalf("writeTimeSeriesData got %q, want to contain %q", got, want)
-	}
-}
